Move stateful variables and methods from TripleAUnit into Unit
