+++ game-core/src/main/java/games/strategy/triplea/TripleAUnit.java
@@ -101,109 +101,7 @@ public class TripleAUnit extends Unit {
     return transportedBy;
   }
 
-  private void setTransportedBy(final TripleAUnit transportedBy) {
-    this.transportedBy = transportedBy;
-  }
 
-  /**
-   * This is a very slow method because it checks all territories on the map. Try not to use this
-   * method if possible.
-   *
-   * @return Unmodifiable collection of units that the given transport is transporting.
-   */
-  public List<Unit> getTransporting() {
-    if (Matches.unitCanTransport().test(this) || Matches.unitIsCarrier().test(this)) {
-      // we don't store the units we are transporting
-      // rather we look at the transported by property of units
-      for (final Territory t : getData().getMap()) {
-        // find the territory this transport is in
-        if (t.getUnitCollection().contains(this)) {
-          return getTransporting(t.getUnitCollection());
-        }
-      }
-    }
-    return List.of();
-  }
-
-  /** @return Unmodifiable collection of units that the given transport is transporting. */
-  public List<Unit> getTransporting(final Collection<Unit> transportedUnitsPossible) {
-    // we don't store the units we are transporting
-    // rather we look at the transported by property of units
-    return Collections.unmodifiableList(
-        CollectionUtils.getMatches(
-            transportedUnitsPossible, o -> equals(get(o).getTransportedBy())));
-  }
-
-  public List<Unit> getUnloaded() {
-    return unloaded;
-  }
-
-  private void setUnloaded(final List<Unit> unloaded) {
-    if (unloaded == null || unloaded.isEmpty()) {
-      this.unloaded = List.of();
-    } else {
-      this.unloaded = new ArrayList<>(unloaded);
-    }
-  }
-
-  public boolean getWasLoadedThisTurn() {
-    return wasLoadedThisTurn;
-  }
-
-  private void setWasLoadedThisTurn(final boolean value) {
-    wasLoadedThisTurn = value;
-  }
-
-  public Territory getUnloadedTo() {
-    return unloadedTo;
-  }
-
-  private void setUnloadedTo(final Territory unloadedTo) {
-    this.unloadedTo = unloadedTo;
-  }
-
-  public Territory getOriginatedFrom() {
-    return originatedFrom;
-  }
-
-  private void setOriginatedFrom(final Territory t) {
-    originatedFrom = t;
-  }
-
-  public boolean getWasUnloadedInCombatPhase() {
-    return wasUnloadedInCombatPhase;
-  }
-
-  private void setWasUnloadedInCombatPhase(final boolean value) {
-    wasUnloadedInCombatPhase = value;
-  }
-
-  public BigDecimal getAlreadyMoved() {
-    return alreadyMoved;
-  }
-
-  public void setAlreadyMoved(final BigDecimal alreadyMoved) {
-    this.alreadyMoved = alreadyMoved;
-  }
-
-  private void setBonusMovement(final int bonusMovement) {
-    this.bonusMovement = bonusMovement;
-  }
-
-  public int getBonusMovement() {
-    return bonusMovement;
-  }
-
-  /** Does not account for any movement already made. Generally equal to UnitType movement */
-  public int getMaxMovementAllowed() {
-    return Math.max(0, bonusMovement + UnitAttachment.get(getType()).getMovement(getOwner()));
-  }
-
-  public BigDecimal getMovementLeft() {
-    return new BigDecimal(UnitAttachment.get(getType()).getMovement(getOwner()))
-        .add(new BigDecimal(bonusMovement))
-        .subtract(alreadyMoved);
-  }
 
   /**
    * Returns a tuple whose first element indicates the minimum movement remaining for the specified
@@ -227,162 +125,6 @@ public class TripleAUnit extends Unit {
       min = max;
     }
     return Tuple.of(min, max);
-  }
-
-  public int getUnitDamage() {
-    return unitDamage;
-  }
-
-  public void setUnitDamage(final int unitDamage) {
-    this.unitDamage = unitDamage;
-  }
-
-  public boolean getSubmerged() {
-    return submerged;
-  }
-
-  public void setSubmerged(final boolean submerged) {
-    this.submerged = submerged;
-  }
-
-  public GamePlayer getOriginalOwner() {
-    return originalOwner;
-  }
-
-  private void setOriginalOwner(final GamePlayer originalOwner) {
-    this.originalOwner = originalOwner;
-  }
-
-  public boolean getWasInCombat() {
-    return wasInCombat;
-  }
-
-  private void setWasInCombat(final boolean value) {
-    wasInCombat = value;
-  }
-
-  public boolean getWasScrambled() {
-    return wasScrambled;
-  }
-
-  private void setWasScrambled(final boolean value) {
-    wasScrambled = value;
-  }
-
-  public int getMaxScrambleCount() {
-    return maxScrambleCount;
-  }
-
-  private void setMaxScrambleCount(final int value) {
-    maxScrambleCount = value;
-  }
-
-  public int getLaunched() {
-    return launched;
-  }
-
-  private void setLaunched(final int value) {
-    launched = value;
-  }
-
-  public boolean getAirborne() {
-    return airborne;
-  }
-
-  private void setAirborne(final boolean value) {
-    airborne = value;
-  }
-
-  public boolean getChargedFlatFuelCost() {
-    return chargedFlatFuelCost;
-  }
-
-  private void setChargedFlatFuelCost(final boolean value) {
-    chargedFlatFuelCost = value;
-  }
-
-  private void setWasInAirBattle(final boolean value) {
-    wasInAirBattle = value;
-  }
-
-  public boolean getWasInAirBattle() {
-    return wasInAirBattle;
-  }
-
-  public boolean getWasLoadedAfterCombat() {
-    return wasLoadedAfterCombat;
-  }
-
-  private void setWasLoadedAfterCombat(final boolean value) {
-    wasLoadedAfterCombat = value;
-  }
-
-  public boolean getWasAmphibious() {
-    return wasAmphibious;
-  }
-
-  private void setWasAmphibious(final boolean value) {
-    wasAmphibious = value;
-  }
-
-  public boolean getDisabled() {
-    return disabled;
-  }
-
-  private void setDisabled(final boolean value) {
-    disabled = value;
-  }
-
-  /**
-   * How much more damage can this unit take? Will return 0 if the unit cannot be damaged, or is at
-   * max damage.
-   */
-  public int getHowMuchMoreDamageCanThisUnitTake(final Unit u, final Territory t) {
-    if (!Matches.unitCanBeDamaged().test(u)) {
-      return 0;
-    }
-    final TripleAUnit taUnit = (TripleAUnit) u;
-    return Properties.getDamageFromBombingDoneToUnitsInsteadOfTerritories(u.getData())
-        ? Math.max(0, getHowMuchDamageCanThisUnitTakeTotal(u, t) - taUnit.getUnitDamage())
-        : Integer.MAX_VALUE;
-  }
-
-  /**
-   * How much damage is the max this unit can take, accounting for territory, etc. Will return -1 if
-   * the unit is of the type that cannot be damaged
-   */
-  public int getHowMuchDamageCanThisUnitTakeTotal(final Unit u, final Territory t) {
-    if (!Matches.unitCanBeDamaged().test(u)) {
-      return -1;
-    }
-    final UnitAttachment ua = UnitAttachment.get(u.getType());
-    final int territoryUnitProduction = TerritoryAttachment.getUnitProduction(t);
-    if (Properties.getDamageFromBombingDoneToUnitsInsteadOfTerritories(u.getData())) {
-      if (ua.getMaxDamage() <= 0) {
-        // factories may or may not have max damage set, so we must still determine here
-        // assume that if maxDamage <= 0, then the max damage must be based on the territory value
-        // can use "production" or "unitProduction"
-        return territoryUnitProduction * 2;
-      }
-
-      if (Matches.unitCanProduceUnits().test(u)) {
-        // can use "production" or "unitProduction"
-        return (ua.getCanProduceXUnits() < 0)
-            ? territoryUnitProduction * ua.getMaxDamage()
-            : ua.getMaxDamage();
-      }
-
-      return ua.getMaxDamage();
-    }
-
-    return Integer.MAX_VALUE;
-  }
-
-  public int getHowMuchCanThisUnitBeRepaired(final Unit u, final Territory t) {
-    return Math.max(
-        0,
-        (this.getHowMuchDamageCanThisUnitTakeTotal(u, t)
-            - this.getHowMuchMoreDamageCanThisUnitTake(u, t)));
   }
 
   public static int getProductionPotentialOfTerritory(
@@ -558,50 +300,6 @@ public class TripleAUnit extends Unit {
   public Map<String, MutableProperty<?>> getPropertyMap() {
     return ImmutableMap.<String, MutableProperty<?>>builder()
         .putAll(super.getPropertyMap())
-        .put(
-            "transportedBy",
-            MutableProperty.ofSimple(this::setTransportedBy, this::getTransportedBy))
-        .put("unloaded", MutableProperty.ofSimple(this::setUnloaded, this::getUnloaded))
-        .put(
-            "wasLoadedThisTurn",
-            MutableProperty.ofSimple(this::setWasLoadedThisTurn, this::getWasLoadedThisTurn))
-        .put("unloadedTo", MutableProperty.ofSimple(this::setUnloadedTo, this::getUnloadedTo))
-        .put(
-            "wasUnloadedInCombatPhase",
-            MutableProperty.ofSimple(
-                this::setWasUnloadedInCombatPhase, this::getWasUnloadedInCombatPhase))
-        .put("alreadyMoved", MutableProperty.ofSimple(this::setAlreadyMoved, this::getAlreadyMoved))
-        .put(
-            "bonusMovement",
-            MutableProperty.ofSimple(this::setBonusMovement, this::getBonusMovement))
-        .put("unitDamage", MutableProperty.ofSimple(this::setUnitDamage, this::getUnitDamage))
-        .put("submerged", MutableProperty.ofSimple(this::setSubmerged, this::getSubmerged))
-        .put(
-            "originalOwner",
-            MutableProperty.ofSimple(this::setOriginalOwner, this::getOriginalOwner))
-        .put("wasInCombat", MutableProperty.ofSimple(this::setWasInCombat, this::getWasInCombat))
-        .put(
-            "wasLoadedAfterCombat",
-            MutableProperty.ofSimple(this::setWasLoadedAfterCombat, this::getWasLoadedAfterCombat))
-        .put(
-            "wasAmphibious",
-            MutableProperty.ofSimple(this::setWasAmphibious, this::getWasAmphibious))
-        .put(
-            "originatedFrom",
-            MutableProperty.ofSimple(this::setOriginatedFrom, this::getOriginatedFrom))
-        .put("wasScrambled", MutableProperty.ofSimple(this::setWasScrambled, this::getWasScrambled))
-        .put(
-            "maxScrambleCount",
-            MutableProperty.ofSimple(this::setMaxScrambleCount, this::getMaxScrambleCount))
-        .put(
-            "wasInAirBattle",
-            MutableProperty.ofSimple(this::setWasInAirBattle, this::getWasInAirBattle))
-        .put("disabled", MutableProperty.ofSimple(this::setDisabled, this::getDisabled))
-        .put("launched", MutableProperty.ofSimple(this::setLaunched, this::getLaunched))
-        .put("airborne", MutableProperty.ofSimple(this::setAirborne, this::getAirborne))
-        .put(
-            "chargedFlatFuelCost",
-            MutableProperty.ofSimple(this::setChargedFlatFuelCost, this::getChargedFlatFuelCost))
         .build();
   }
 }
